requirements.txt
playwright>=1.40.0
openai>=1.10.0
python-dotenv>=1.0.0
pillow>=10.1.0
click>=8.1.0
pypdf>=3.17.0
python-docx>=1.1.0
